Slight refactoring.
